+++ bacnet-stack/wp.c
@@ -88,7 +88,7 @@ int wp_decode_service_request(
   uint32_t len_value_type = 0;
   int type = 0; // for decoding
   int property = 0; // for decoding
-  unsigned unsigned_value = 0;
+  uint32_t unsigned_value = 0;
 
   // check for value pointers
   if (apdu_len && data)
@@ -143,7 +143,8 @@ int wp_decode_service_request(
       {
         len += tag_len;
         len = decode_unsigned(&apdu[len], len_value_type, &unsigned_value);
-        data->priority = unsigned_value;
+		/* FIXME: bounds check */
+        data->priority = (uint8_t)unsigned_value;
       }
     }
   }
